# api/openapi.yaml
openapi: 3.0.3
info:
  title: Balemoh API
  version: 0.1.0
paths:
  /healthz:
    get:
      operationId: GetHealthz
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Service is unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/staging/services:
    get:
      operationId: GetStagingServices
      responses:
        '200':
          description: Discovered service candidates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceListResponse'
        '503':
          description: Catalog is unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/staging/services/{serviceId}/pin:
    parameters:
      - name: serviceId
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: PinStagingService
      responses:
        '200':
          description: Updated pinned service candidate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceCandidate'
        '404':
          description: Candidate was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Catalog is unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: UnpinStagingService
      responses:
        '200':
          description: Updated unpinned service candidate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceCandidate'
        '404':
          description: Candidate was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Catalog is unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/homepage/services:
    get:
      operationId: GetHomepageServices
      responses:
        '200':
          description: Pinned homepage services
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceListResponse'
        '503':
          description: Catalog is unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/discovery/sync:
    post:
      operationId: SyncDiscovery
      responses:
        '200':
          description: Discovery synchronization result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoverySyncResponse'
        '503':
          description: Discovery synchronization failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/federation/snapshots:
    post:
      operationId: ImportFederationSnapshot
      security:
        - FederationBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FederationSnapshot'
      responses:
        '200':
          description: Federated discovery snapshot imported
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoverySyncResponse'
        '400':
          description: Snapshot is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Federation credentials are invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Source is not registered at this gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Federation ingestion is not enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Snapshot could not be stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    FederationBearer:
      type: http
      scheme: bearer
  schemas:
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ok
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    SourceRef:
      type: object
      required:
        - kind
        - id
      properties:
        kind:
          type: string
        id:
          type: string
    ResourceRef:
      type: object
      required:
        - kind
        - name
      properties:
        kind:
          type: string
        namespace:
          type: string
        name:
          type: string
    ServiceEndpoint:
      type: object
      required:
        - name
        - url
        - port
        - protocol
        - provenance
      properties:
        name:
          type: string
        url:
          type: string
          description: Empty when discovery has no exact hostname or route; HTTPRoute observations may be scheme-relative.
        port:
          type: integer
          format: int32
          minimum: 0
          maximum: 65535
          description: Zero when no port was observed.
        protocol:
          type: string
        provenance:
          type: string
          description: Adapter evidence for the endpoint observation.
    ServiceCandidate:
      type: object
      required:
        - id
        - source
        - resource
        - displayName
        - description
        - metadata
        - endpoints
        - images
        - observedAt
        - pinned
        - pinnedAt
      properties:
        id:
          type: string
        source:
          $ref: '#/components/schemas/SourceRef'
        resource:
          $ref: '#/components/schemas/ResourceRef'
        displayName:
          type: string
        description:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
        endpoints:
          type: array
          items:
            $ref: '#/components/schemas/ServiceEndpoint'
        images:
          type: array
          description: Container images observed for the candidate, when available.
          items:
            type: string
        observedAt:
          type: string
          format: date-time
        pinned:
          type: boolean
        pinnedAt:
          type: string
          format: date-time
          nullable: true
    ServiceListResponse:
      type: object
      required:
        - services
      properties:
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceCandidate'
    DiscoverySyncResponse:
      type: object
      required:
        - sources
        - candidates
      properties:
        sources:
          type: integer
          format: int32
          minimum: 0
        candidates:
          type: integer
          format: int32
          minimum: 0
    FederationSnapshot:
      type: object
      required:
        - source
        - candidates
        - observedAt
      properties:
        source:
          $ref: '#/components/schemas/SourceRef'
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/FederatedCandidate'
        observedAt:
          type: string
          format: date-time
    FederatedCandidate:
      type: object
      required:
        - id
        - source
        - resource
        - displayName
        - description
        - metadata
        - endpoints
        - images
        - observedAt
      properties:
        id:
          type: string
        source:
          $ref: '#/components/schemas/SourceRef'
        resource:
          $ref: '#/components/schemas/ResourceRef'
        displayName:
          type: string
        description:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
        endpoints:
          type: array
          items:
            $ref: '#/components/schemas/ServiceEndpoint'
        images:
          type: array
          items:
            type: string
        observedAt:
          type: string
          format: date-time