# internal/api/generated/balemoh.gen.go
//go:build go1.22

// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.2 DO NOT EDIT.
package generated

import (
	"context"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

const (
	FederationBearerScopes federationBearerContextKey = "FederationBearer.Scopes"
)

// Defines values for HealthResponseStatus.
const (
	Ok HealthResponseStatus = "ok"
)

// Valid indicates whether the value is a known member of the HealthResponseStatus enum.
func (e HealthResponseStatus) Valid() bool {
	switch e {
	case Ok:
		return true
	default:
		return false
	}
}

// DiscoverySyncResponse defines model for DiscoverySyncResponse.
type DiscoverySyncResponse struct {
	Candidates int32 `json:"candidates"`
	Sources    int32 `json:"sources"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Code    string `json:"code"`
	Message string `json:"message"`
}

// FederatedCandidate defines model for FederatedCandidate.
type FederatedCandidate struct {
	Description string            `json:"description"`
	DisplayName string            `json:"displayName"`
	Endpoints   []ServiceEndpoint `json:"endpoints"`
	Id          string            `json:"id"`
	Images      []string          `json:"images"`
	Metadata    map[string]string `json:"metadata"`
	ObservedAt  time.Time         `json:"observedAt"`
	Resource    ResourceRef       `json:"resource"`
	Source      SourceRef         `json:"source"`
}

// FederationSnapshot defines model for FederationSnapshot.
type FederationSnapshot struct {
	Candidates []FederatedCandidate `json:"candidates"`
	ObservedAt time.Time            `json:"observedAt"`
	Source     SourceRef            `json:"source"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Status HealthResponseStatus `json:"status"`
}

// HealthResponseStatus defines model for HealthResponse.Status.
type HealthResponseStatus string

// ResourceRef defines model for ResourceRef.
type ResourceRef struct {
	Kind      string  `json:"kind"`
	Name      string  `json:"name"`
	Namespace *string `json:"namespace,omitempty"`
}

// ServiceCandidate defines model for ServiceCandidate.
type ServiceCandidate struct {
	Description string            `json:"description"`
	DisplayName string            `json:"displayName"`
	Endpoints   []ServiceEndpoint `json:"endpoints"`
	Id          string            `json:"id"`

	// Images Container images observed for the candidate, when available.
	Images     []string          `json:"images"`
	Metadata   map[string]string `json:"metadata"`
	ObservedAt time.Time         `json:"observedAt"`
	Pinned     bool              `json:"pinned"`
	PinnedAt   *time.Time        `json:"pinnedAt"`
	Resource   ResourceRef       `json:"resource"`
	Source     SourceRef         `json:"source"`
}

// ServiceEndpoint defines model for ServiceEndpoint.
type ServiceEndpoint struct {
	Name string `json:"name"`

	// Port Zero when no port was observed.
	Port     int32  `json:"port"`
	Protocol string `json:"protocol"`

	// Provenance Adapter evidence for the endpoint observation.
	Provenance string `json:"provenance"`

	// Url Empty when discovery has no exact hostname or route; HTTPRoute observations may be scheme-relative.
	Url string `json:"url"`
}

// ServiceListResponse defines model for ServiceListResponse.
type ServiceListResponse struct {
	Services []ServiceCandidate `json:"services"`
}

// SourceRef defines model for SourceRef.
type SourceRef struct {
	Id   string `json:"id"`
	Kind string `json:"kind"`
}

// federationBearerContextKey is the context key for FederationBearer security scheme
type federationBearerContextKey string

// ImportFederationSnapshotJSONRequestBody defines body for ImportFederationSnapshot for application/json ContentType.
type ImportFederationSnapshotJSONRequestBody = FederationSnapshot

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (POST /api/v1/discovery/sync)
	SyncDiscovery(w http.ResponseWriter, r *http.Request)

	// (POST /api/v1/federation/snapshots)
	ImportFederationSnapshot(w http.ResponseWriter, r *http.Request)

	// (GET /api/v1/homepage/services)
	GetHomepageServices(w http.ResponseWriter, r *http.Request)

	// (GET /api/v1/staging/services)
	GetStagingServices(w http.ResponseWriter, r *http.Request)

	// (DELETE /api/v1/staging/services/{serviceId}/pin)
	UnpinStagingService(w http.ResponseWriter, r *http.Request, serviceId string)

	// (POST /api/v1/staging/services/{serviceId}/pin)
	PinStagingService(w http.ResponseWriter, r *http.Request, serviceId string)

	// (GET /healthz)
	GetHealthz(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// SyncDiscovery operation middleware
func (siw *ServerInterfaceWrapper) SyncDiscovery(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SyncDiscovery(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ImportFederationSnapshot operation middleware
func (siw *ServerInterfaceWrapper) ImportFederationSnapshot(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, FederationBearerScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ImportFederationSnapshot(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetHomepageServices operation middleware
func (siw *ServerInterfaceWrapper) GetHomepageServices(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHomepageServices(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetStagingServices operation middleware
func (siw *ServerInterfaceWrapper) GetStagingServices(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetStagingServices(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UnpinStagingService operation middleware
func (siw *ServerInterfaceWrapper) UnpinStagingService(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "serviceId" -------------
	var serviceId string

	err = runtime.BindStyledParameterWithOptions("simple", "serviceId", r.PathValue("serviceId"), &serviceId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "serviceId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UnpinStagingService(w, r, serviceId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PinStagingService operation middleware
func (siw *ServerInterfaceWrapper) PinStagingService(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "serviceId" -------------
	var serviceId string

	err = runtime.BindStyledParameterWithOptions("simple", "serviceId", r.PathValue("serviceId"), &serviceId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "serviceId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PinStagingService(w, r, serviceId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetHealthz operation middleware
func (siw *ServerInterfaceWrapper) GetHealthz(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealthz(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/v1/discovery/sync", wrapper.SyncDiscovery)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/v1/federation/snapshots", wrapper.ImportFederationSnapshot)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/v1/homepage/services", wrapper.GetHomepageServices)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/v1/staging/services", wrapper.GetStagingServices)
	m.HandleFunc(http.MethodDelete+" "+options.BaseURL+"/api/v1/staging/services/{serviceId}/pin", wrapper.UnpinStagingService)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/v1/staging/services/{serviceId}/pin", wrapper.PinStagingService)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/healthz", wrapper.GetHealthz)

	return m
}